# Modify the basic structure `NewInfoElementWithValue`