--- a/pkg/entities/ie.go
+++ b/pkg/entities/ie.go
@@ -93,13 +93,6 @@ type InfoElement struct {
 	Len uint16
 }
 
-// InfoElementWithValue represents mapping from element to value for data records
-type InfoElementWithValue struct {
-	Element *InfoElement
-	Value   interface{}
-	Length  int
-}
-
 func NewInfoElement(name string, ieID uint16, ieType IEDataType, entID uint32, len uint16) *InfoElement {
 	return &InfoElement{
 		Name:         name,
@@ -107,12 +100,6 @@ func NewInfoElement(name string, ieID uint16, ieType IEDataType, entID uint32, l
 		DataType:     ieType,
 		EnterpriseId: entID,
 		Len:          len,
-	}
-}
-
-func NewInfoElementWithValue(element *InfoElement, value interface{}) InfoElementWithValue {
-	return InfoElementWithValue{
-		element, value, 0,
 	}
 }
 
@@ -172,8 +159,8 @@ func IsValidDataType(tp IEDataType) bool {
 	return tp != InvalidDataType
 }
 
-// DecodeToIEDataType is to decode to specific type
-func DecodeToIEDataType(dataType IEDataType, val interface{}) (interface{}, error) {
+// decodeToIEDataType is to decode to specific type. This is only used for testing.
+func decodeToIEDataType(dataType IEDataType, val interface{}) (interface{}, error) {
 	value, ok := val.([]byte)
 	if !ok {
 		return nil, fmt.Errorf("error when converting value to []bytes for decoding")
@@ -224,7 +211,135 @@ func DecodeToIEDataType(dataType IEDataType, val interface{}) (interface{}, erro
 	}
 }
 
-// EncodeToIEDataType is to encode data to specific type to the buff
+// DecodeAndCreateInfoElementWithValue takes in the info element and its value in bytes, and
+// returns appropriate InfoElementWithValue.
+func DecodeAndCreateInfoElementWithValue(element *InfoElement, value []byte) (InfoElementWithValue, error) {
+	switch element.DataType {
+	case Unsigned8:
+		var val uint8
+		if value == nil {
+			val = 0
+		} else {
+			val = value[0]
+		}
+		return NewUnsigned8InfoElement(element, val), nil
+	case Unsigned16:
+		var val uint16
+		if value == nil {
+			val = 0
+		} else {
+			val = binary.BigEndian.Uint16(value)
+		}
+		return NewUnsigned16InfoElement(element, val), nil
+	case Unsigned32:
+		var val uint32
+		if value == nil {
+			val = 0
+		} else {
+			val = binary.BigEndian.Uint32(value)
+		}
+		return NewUnsigned32InfoElement(element, val), nil
+	case Unsigned64:
+		var val uint64
+		if value == nil {
+			val = 0
+		} else {
+			val = binary.BigEndian.Uint64(value)
+		}
+		return NewUnsigned64InfoElement(element, val), nil
+	case Signed8:
+		var val int8
+		if value == nil {
+			val = 0
+		} else {
+			val = int8(value[0])
+		}
+		return NewSigned8InfoElement(element, val), nil
+	case Signed16:
+		var val int16
+		if value == nil {
+			val = 0
+		} else {
+			val = int16(binary.BigEndian.Uint16(value))
+		}
+		return NewSigned16InfoElement(element, val), nil
+	case Signed32:
+		var val int32
+		if value == nil {
+			val = 0
+		} else {
+			val = int32(binary.BigEndian.Uint32(value))
+		}
+		return NewSigned32InfoElement(element, val), nil
+	case Signed64:
+		var val int64
+		if value == nil {
+			val = 0
+		}
+		val = int64(binary.BigEndian.Uint64(value))
+		return NewSigned64InfoElement(element, val), nil
+	case Float32:
+		var val float32
+		if value == nil {
+			val = 0
+		} else {
+			val = math.Float32frombits(binary.BigEndian.Uint32(value))
+		}
+		return NewFloat32InfoElement(element, val), nil
+	case Float64:
+		var val float64
+		if value == nil {
+			val = 0
+		} else {
+			val = math.Float64frombits(binary.BigEndian.Uint64(value))
+		}
+		return NewFloat64InfoElement(element, val), nil
+	case Boolean:
+		if value == nil {
+			return NewBoolInfoElement(element, false), nil
+		}
+		if int8(value[0]) == 1 {
+			return NewBoolInfoElement(element, true), nil
+		} else {
+			return NewBoolInfoElement(element, false), nil
+		}
+	case DateTimeSeconds:
+		var val uint32
+		if value == nil {
+			val = 0
+		} else {
+			val = binary.BigEndian.Uint32(value)
+		}
+		return NewDateTimeSecondsInfoElement(element, val), nil
+	case DateTimeMilliseconds:
+		var val uint64
+		if value == nil {
+			val = 0
+		} else {
+			val = binary.BigEndian.Uint64(value)
+		}
+		return NewDateTimeMillisecondsInfoElement(element, val), nil
+	case DateTimeMicroseconds, DateTimeNanoseconds:
+		return nil, fmt.Errorf("API does not support micro and nano seconds types yet")
+	case MacAddress:
+		return NewMacAddressInfoElement(element, value), nil
+	case Ipv4Address, Ipv6Address:
+		return NewIPAddressInfoElement(element, value), nil
+	case String:
+		var val string
+		if value == nil {
+			val = ""
+		} else {
+			val = string(value)
+		}
+		return NewStringInfoElement(element, val), nil
+	default:
+		return nil, fmt.Errorf("API supports only valid information elements with datatypes given in RFC7011")
+	}
+}
+
+// EncodeToIEDataType is to encode data to specific type to the buff. This is only
+// used for testing.
 func EncodeToIEDataType(dataType IEDataType, val interface{}) ([]byte, error) {
 	switch dataType {
 	case Unsigned8:
@@ -390,76 +505,77 @@ func EncodeToIEDataType(dataType IEDataType, val interface{}) ([]byte, error) {
 	return nil, fmt.Errorf("API supports only valid information elements with datatypes given in RFC7011")
 }
 
-// encodeToBuff is to encode data to specific type to the buff
-func encodeToBuff(dataType IEDataType, val interface{}, length int, buffer []byte, index int) error {
-	if index+length > len(buffer) {
+// encodeInfoElementValueToBuff is to encode data to specific type to the buff
+func encodeInfoElementValueToBuff(element InfoElementWithValue, buffer []byte, index int) error {
+	infoElem := element.GetInfoElement()
+	if index+element.GetLength() > len(buffer) {
 		return fmt.Errorf("buffer size is not enough for encoding")
 	}
-	switch dataType {
+	switch infoElem.DataType {
 	case Unsigned8:
-		v, ok := val.(uint8)
-		if !ok {
-			return fmt.Errorf("val argument %v is not of type uint8", val)
+		v, err := element.GetUnsigned8Value()
+		if err != nil {
+			return err
 		}
 		copy(buffer[index:index+1], []byte{v})
 	case Unsigned16:
-		v, ok := val.(uint16)
-		if !ok {
-			return fmt.Errorf("val argument %v is not of type uint16", val)
+		v, err := element.GetUnsigned16Value()
+		if err != nil {
+			return err
 		}
 		binary.BigEndian.PutUint16(buffer[index:], v)
 	case Unsigned32:
-		v, ok := val.(uint32)
-		if !ok {
-			return fmt.Errorf("val argument %v is not of type uint32", val)
+		v, err := element.GetUnsigned32Value()
+		if err != nil {
+			return err
 		}
 		binary.BigEndian.PutUint32(buffer[index:], v)
 	case Unsigned64:
-		v, ok := val.(uint64)
-		if !ok {
-			return fmt.Errorf("val argument %v is not of type uint64", val)
+		v, err := element.GetUnsigned64Value()
+		if err != nil {
+			return err
 		}
 		binary.BigEndian.PutUint64(buffer[index:], v)
 	case Signed8:
-		v, ok := val.(int8)
-		if !ok {
-			return fmt.Errorf("val argument %v is not of type int8", val)
+		v, err := element.GetSigned8Value()
+		if err != nil {
+			return err
 		}
 		copy(buffer[index:index+1], []byte{byte(v)})
 	case Signed16:
-		v, ok := val.(int16)
-		if !ok {
-			return fmt.Errorf("val argument %v is not of type int16", val)
+		v, err := element.GetSigned16Value()
+		if err != nil {
+			return err
 		}
 		binary.BigEndian.PutUint16(buffer[index:], uint16(v))
 	case Signed32:
-		v, ok := val.(int32)
-		if !ok {
-			return fmt.Errorf("val argument %v is not of type int32", val)
+		v, err := element.GetSigned32Value()
+		if err != nil {
+			return err
 		}
 		binary.BigEndian.PutUint32(buffer[index:], uint32(v))
 	case Signed64:
-		v, ok := val.(int64)
-		if !ok {
-			return fmt.Errorf("val argument %v is not of type int64", val)
+		v, err := element.GetSigned64Value()
+		if err != nil {
+			return err
 		}
 		binary.BigEndian.PutUint64(buffer[index:], uint64(v))
 	case Float32:
-		v, ok := val.(float32)
-		if !ok {
-			return fmt.Errorf("val argument %v is not of type float32", val)
+		v, err := element.GetFloat32Value()
+		if err != nil {
+			return err
 		}
 		binary.BigEndian.PutUint32(buffer[index:], math.Float32bits(v))
 	case Float64:
-		v, ok := val.(float64)
-		if !ok {
-			return fmt.Errorf("val argument %v is not of type float64", val)
+		v, err := element.GetFloat64Value()
+		if err != nil {
+			return err
 		}
 		binary.BigEndian.PutUint64(buffer[index:], math.Float64bits(v))
 	case Boolean:
-		v, ok := val.(bool)
-		if !ok {
-			return fmt.Errorf("val argument %v is not of type bool", val)
+		v, err := element.GetBooleanValue()
+		if err != nil {
+			return err
 		}
 		// Following boolean spec from RFC7011
 		indicator := byte(int8(1))
@@ -468,15 +584,15 @@ func encodeToBuff(dataType IEDataType, val interface{}, length int, buffer []byt
 		}
 		copy(buffer[index:index+1], []byte{indicator})
 	case DateTimeSeconds:
-		v, ok := val.(uint32)
-		if !ok {
-			return fmt.Errorf("val argument %v is not of type uint32", val)
+		v, err := element.GetUnsigned32Value()
+		if err != nil {
+			return err
 		}
 		binary.BigEndian.PutUint32(buffer[index:], v)
 	case DateTimeMilliseconds:
-		v, ok := val.(uint64)
-		if !ok {
-			return fmt.Errorf("val argument %v is not of type uint64", val)
+		v, err := element.GetUnsigned64Value()
+		if err != nil {
+			return err
 		}
 		binary.BigEndian.PutUint64(buffer[index:], v)
 		// Currently only supporting seconds and milliseconds
@@ -485,16 +601,16 @@ func encodeToBuff(dataType IEDataType, val interface{}, length int, buffer []byt
 		return fmt.Errorf("API does not support micro and nano seconds types yet")
 	case MacAddress:
 		// Expects net.Hardware type
-		v, ok := val.(net.HardwareAddr)
-		if !ok {
-			return fmt.Errorf("val argument %v is not of type net.HardwareAddr for this element", val)
+		v, err := element.GetMacAddressValue()
+		if err != nil {
+			return err
 		}
 		copy(buffer[index:], v)
 	case Ipv4Address:
 		// Expects net.IP type
-		v, ok := val.(net.IP)
-		if !ok {
-			return fmt.Errorf("val argument %v is not of type net.IP for this element", val)
+		v, err := element.GetIPAddressValue()
+		if err != nil {
+			return err
 		}
 		if ipv4Add := v.To4(); ipv4Add != nil {
 			copy(buffer[index:], ipv4Add)
@@ -503,9 +619,9 @@ func encodeToBuff(dataType IEDataType, val interface{}, length int, buffer []byt
 		}
 	case Ipv6Address:
 		// Expects net.IP type
-		v, ok := val.(net.IP)
-		if !ok {
-			return fmt.Errorf("val argument %v is not of type net.IP for this element", val)
+		v, err := element.GetIPAddressValue()
+		if err != nil {
+			return err
 		}
 		if ipv6Add := v.To16(); ipv6Add != nil {
 			copy(buffer[index:], ipv6Add)
@@ -513,9 +629,9 @@ func encodeToBuff(dataType IEDataType, val interface{}, length int, buffer []byt
 			return fmt.Errorf("provided IPv6 address %v is not of correct length", v)
 		}
 	case String:
-		v, ok := val.(string)
-		if !ok {
-			return fmt.Errorf("val argument %v is not of type string for this element", val)
+		v, err := element.GetStringValue()
+		if err != nil {
+			return err
 		}
 		if len(v) < 255 {
 			buffer[index] = uint8(len(v))
@@ -533,17 +649,4 @@ func encodeToBuff(dataType IEDataType, val interface{}, length int, buffer []byt
 		return fmt.Errorf("API supports only valid information elements with datatypes given in RFC7011")
 	}
 	return nil
-}
-
-func setInfoElementLen(element *InfoElementWithValue) {
-	if element.Element.DataType != String {
-		element.Length = int(element.Element.Len)
-	} else {
-		v := element.Value.(string)
-		if len(v) < 255 {
-			element.Length = len(v) + 1
-		} else {
-			element.Length = len(v) + 3
-		}
-	}
 }