--- a/pkg/entities/record.go
+++ b/pkg/entities/record.go
@@ -31,14 +31,13 @@ import (
 
 type Record interface {
 	PrepareRecord() error
-	AddInfoElement(element *InfoElementWithValue) error
+	AddInfoElement(element InfoElementWithValue) error
 	// TODO: Functions for multiple elements as well.
 	GetBuffer() []byte
 	GetTemplateID() uint16
 	GetFieldCount() uint16
 	GetOrderedElementList() []InfoElementWithValue
-	GetInfoElementWithValue(name string) (*InfoElementWithValue, int, bool)
-	SetInfoElementWithValue(index int, infoElement InfoElementWithValue) error
+	GetInfoElementWithValue(name string) (InfoElementWithValue, int, bool)
 	GetRecordLength() int
 	GetMinDataRecordLen() uint16
 }
@@ -103,22 +102,13 @@ func (b *baseRecord) GetOrderedElementList() []InfoElementWithValue {
 	return b.orderedElementList
 }
 
-func (b *baseRecord) GetInfoElementWithValue(name string) (*InfoElementWithValue, int, bool) {
-	for i := range b.orderedElementList {
-		element := &b.orderedElementList[i]
-		if element.Element.Name == name {
+func (b *baseRecord) GetInfoElementWithValue(name string) (InfoElementWithValue, int, bool) {
+	for i, element := range b.orderedElementList {
+		if element.GetInfoElement().Name == name {
 			return element, i, true
 		}
 	}
 	return nil, 0, false
-}
-
-func (b *baseRecord) SetInfoElementWithValue(index int, infoElement InfoElementWithValue) error {
-	if b.orderedElementList[index].Element.Name != infoElement.Element.Name {
-		return fmt.Errorf("element not present in the record")
-	}
-	b.orderedElementList[index] = infoElement
-	return nil
 }
 
 func (d *dataRecord) PrepareRecord() error {
@@ -133,11 +123,11 @@ func (d *dataRecord) GetBuffer() []byte {
 	d.buffer = make([]byte, d.len)
 	index := 0
 	for _, element := range d.orderedElementList {
-		err := encodeToBuff(element.Element.DataType, element.Value, element.Length, d.buffer, index)
+		err := encodeInfoElementValueToBuff(element, d.buffer, index)
 		if err != nil {
 			klog.Error(err)
 		}
-		index += element.Length
+		index += element.GetLength()
 	}
 	return d.buffer
 }
@@ -146,23 +136,14 @@ func (d *dataRecord) GetRecordLength() int {
 	return d.len
 }
 
-func (d *dataRecord) AddInfoElement(element *InfoElementWithValue) error {
-	var value interface{}
-	var err error
-	if d.isDecoding {
-		value, err = DecodeToIEDataType(element.Element.DataType, element.Value)
-		if err != nil {
-			return err
-		}
-		element.Value = value
-	} else {
-		setInfoElementLen(element)
-		d.len += element.Length
+func (d *dataRecord) AddInfoElement(element InfoElementWithValue) error {
+	if !d.isDecoding {
+		d.len = d.len + element.GetLength()
 	}
 	if len(d.orderedElementList) <= int(d.fieldCount) {
-		d.orderedElementList = append(d.orderedElementList, *element)
+		d.orderedElementList = append(d.orderedElementList, element)
 	} else {
-		d.orderedElementList[d.fieldCount] = *element
+		d.orderedElementList[d.fieldCount] = element
 	}
 	d.fieldCount++
 	return nil
@@ -175,31 +156,32 @@ func (t *templateRecord) PrepareRecord() error {
 	return nil
 }
 
-func (t *templateRecord) AddInfoElement(element *InfoElementWithValue) error {
+func (t *templateRecord) AddInfoElement(element InfoElementWithValue) error {
+	infoElement := element.GetInfoElement()
 	// val could be used to specify smaller length than default? For now assert it to be nil
-	if element.Value != nil {
-		return fmt.Errorf("AddInfoElement(templateRecord) cannot take value %v (nil is expected)", element.Value)
+	if !element.IsValueEmpty() {
+		return fmt.Errorf("template record cannot take element %v with non-empty value", infoElement.Name)
 	}
 	initialLength := len(t.buffer)
 	// Add field specifier {elementID: uint16, elementLen: uint16}
 	addBytes := make([]byte, 4)
-	binary.BigEndian.PutUint16(addBytes[0:2], element.Element.ElementId)
-	binary.BigEndian.PutUint16(addBytes[2:4], element.Element.Len)
+	binary.BigEndian.PutUint16(addBytes[0:2], infoElement.ElementId)
+	binary.BigEndian.PutUint16(addBytes[2:4], infoElement.Len)
 	t.buffer = append(t.buffer, addBytes...)
-	if element.Element.EnterpriseId != 0 {
+	if infoElement.EnterpriseId != 0 {
 		// Set the MSB of elementID to 1 as per RFC7011
 		t.buffer[initialLength] = t.buffer[initialLength] | 0x80
 		addBytes = make([]byte, 4)
-		binary.BigEndian.PutUint32(addBytes, element.Element.EnterpriseId)
+		binary.BigEndian.PutUint32(addBytes, infoElement.EnterpriseId)
 		t.buffer = append(t.buffer, addBytes...)
 	}
-	t.orderedElementList[t.index] = *element
+	t.orderedElementList[t.index] = element
 	t.index++
 	// Keep track of minimum data record length required for sanity check
-	if element.Element.Len == VariableLength {
+	if infoElement.Len == VariableLength {
 		t.minDataRecLength = t.minDataRecLength + 1
 	} else {
-		t.minDataRecLength = t.minDataRecLength + element.Element.Len
+		t.minDataRecLength = t.minDataRecLength + infoElement.Len
 	}
 	return nil
 }