--- a/pkg/entities/testing/mock_record.go
+++ b/pkg/entities/testing/mock_record.go
@@ -48,7 +48,7 @@ func (m *MockRecord) EXPECT() *MockRecordMockRecorder {
 }
 
 // AddInfoElement mocks base method
-func (m *MockRecord) AddInfoElement(arg0 *entities.InfoElementWithValue) error {
+func (m *MockRecord) AddInfoElement(arg0 entities.InfoElementWithValue) error {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "AddInfoElement", arg0)
 	ret0, _ := ret[0].(error)
@@ -90,10 +90,10 @@ func (mr *MockRecordMockRecorder) GetFieldCount() *gomock.Call {
 }
 
 // GetInfoElementWithValue mocks base method
-func (m *MockRecord) GetInfoElementWithValue(arg0 string) (*entities.InfoElementWithValue, int, bool) {
+func (m *MockRecord) GetInfoElementWithValue(arg0 string) (entities.InfoElementWithValue, int, bool) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "GetInfoElementWithValue", arg0)
-	ret0, _ := ret[0].(*entities.InfoElementWithValue)
+	ret0, _ := ret[0].(entities.InfoElementWithValue)
 	ret1, _ := ret[1].(int)
 	ret2, _ := ret[2].(bool)
 	return ret0, ret1, ret2
@@ -173,18 +173,4 @@ func (m *MockRecord) PrepareRecord() error {
 func (mr *MockRecordMockRecorder) PrepareRecord() *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "PrepareRecord", reflect.TypeOf((*MockRecord)(nil).PrepareRecord))
-}
-
-// SetInfoElementWithValue mocks base method
-func (m *MockRecord) SetInfoElementWithValue(arg0 int, arg1 entities.InfoElementWithValue) error {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "SetInfoElementWithValue", arg0, arg1)
-	ret0, _ := ret[0].(error)
-	return ret0
-}
-
-// SetInfoElementWithValue indicates an expected call of SetInfoElementWithValue
-func (mr *MockRecordMockRecorder) SetInfoElementWithValue(arg0, arg1 interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "SetInfoElementWithValue", reflect.TypeOf((*MockRecord)(nil).SetInfoElementWithValue), arg0, arg1)
 }