Instead of relying on the fact that any data type can be interpreted
through interface{} type using their pointer, we move to InfoElementWithValue
interface. For each data type, we implement this interface with a separate struct.

This results in a very big file pkg/entities/ie_value.go
As there is no good support for generic data types in golang, we need to resort to
this.

Fixes one of the tasks in Issue #209 

In the process, we reduced allocations and memory consumed in the performance
test.

Before fix:
```
BenchmarkMultipleExportersToCollector-16    	       1	1392006670 ns/op	347268992 B/op	 5296157 allocs/op
PASS
ok  	github.com/vmware/go-ipfix/pkg/test	1.829s
```
After fix:
```
BenchmarkMultipleExportersToCollector-16    	       1	1346594175 ns/op	237156360 B/op	 4880056 allocs/op
PASS
ok  	github.com/vmware/go-ipfix/pkg/test	1.865s
```