@srikartati 2021-09-02
Allocated objects and allocated space in `decodeDataSet` is very high and this contributed to very high memory consumption. This is the main method in the collecting process.
By zooming into allocated space for that method, we can see that cumulative allocated space has gone down by almost 50%.
This will greatly reduce the memory fragmentation issue seen in scale tests.

Before fix:
```
(pprof) list decodeDataSet
Total: 352.93MB
ROUTINE ======================== github.com/vmware/go-ipfix/pkg/collector.(*CollectingProcess).decodeDataSet in /Users/stati/work/antrea_all/go-ipfix/pkg/collector/process.go
   59.02MB   116.54MB (flat, cum) 33.02% of Total
         .          .    276:	// make sure template exists
         .          .    277:	template, err := cp.getTemplate(obsDomainID, templateID)
         .          .    278:	if err != nil {
         .          .    279:		return nil, fmt.Errorf("template %d with obsDomainID %d does not exist", templateID, obsDomainID)
         .          .    280:	}
         .        6MB    281:	dataSet := entities.NewSet(true)
         .          .    282:	if err = dataSet.PrepareSet(entities.Data, templateID); err != nil {
         .          .    283:		return nil, err
         .          .    284:	}
         .          .    285:
         .          .    286:	for dataBuffer.Len() > 0 {
   29.52MB    29.52MB    287:		elements := make([]entities.InfoElementWithValue, len(template))
         .          .    288:		for i, element := range template {
         .          .    289:			var length int
         .          .    290:			if element.Len == entities.VariableLength { // string
         .          .    291:				length = getFieldLength(dataBuffer)
         .          .    292:			} else {
         .          .    293:				length = int(element.Len)
         .          .    294:			}
   29.50MB    29.50MB    295:			elements[i] = entities.NewInfoElementWithValue(element, dataBuffer.Next(length))
         .          .    296:		}
         .    51.52MB    297:		err = dataSet.AddRecordWithExtraElements(elements, cp.numExtraElements, templateID)
         .          .    298:		if err != nil {
         .          .    299:			return nil, err
         .          .    300:		}
         .          .    301:	}
         .          .    302:	return dataSet, nil
(pprof)  
```

After fix:

```
(pprof) list decodeDataSet
Total: 228.86MB
ROUTINE ======================== github.com/vmware/go-ipfix/pkg/collector.(*CollectingProcess).decodeDataSet in /Users/stati/work/antrea_all/go-ipfix/pkg/collector/process.go
   11.50MB    61.51MB (flat, cum) 26.88% of Total
         .          .    278:	// make sure template exists
         .          .    279:	template, err := cp.getTemplate(obsDomainID, templateID)
         .          .    280:	if err != nil {
         .          .    281:		return nil, fmt.Errorf("template %d with obsDomainID %d does not exist", templateID, obsDomainID)
         .          .    282:	}
         .        4MB    283:	dataSet := entities.NewSet(true)
         .          .    284:	if err = dataSet.PrepareSet(entities.Data, templateID); err != nil {
         .          .    285:		return nil, err
         .          .    286:	}
         .          .    287:
         .          .    288:	for dataBuffer.Len() > 0 {
   11.50MB    11.50MB    289:		elements := make([]entities.InfoElementWithValue, len(template))
         .          .    290:		for i, element := range template {
         .          .    291:			var length int
         .          .    292:			if element.Len == entities.VariableLength { // string
         .          .    293:				length = getFieldLength(dataBuffer)
         .          .    294:			} else {
         .          .    295:				length = int(element.Len)
         .          .    296:			}
         .    24.50MB    297:			if elements[i], err = entities.GetInfoElementWithValue(element, dataBuffer.Next(length)); err != nil {
         .          .    298:				return nil, err
         .          .    299:			}
         .          .    300:		}
         .    21.50MB    301:		err = dataSet.AddRecordWithExtraElements(elements, cp.numExtraElements, templateID)
         .          .    302:		if err != nil {
         .          .    303:			return nil, err
         .          .    304:		}
         .          .    305:	}
         .          .    306:	return dataSet, nil
(pprof) exit
```