@srikartati 2021-09-03
Ran this performance unit test: https://github.com/antrea-io/antrea/blob/main/pkg/agent/flowexporter/exporter/exporter_perf_test.go#L72

Before fix:
```
 go test -test.v -run=BenchmarkExport -test.benchmem -bench=BenchmarkExportConntrackConns -memprofile memprofile.out -cpuprofile profile.out
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
BenchmarkExportConntrackConns-2   	      96	  16181741 ns/op	 1124179 B/op	   10519 allocs/op
PASS
ok  	antrea.io/antrea/pkg/agent/flowexporter/exporter	9.123s
```
After fix:
```
go test -test.v -run=BenchmarkExport -test.benchmem -bench=BenchmarkExportConntrackConns -memprofile memprofile.out -cpuprofile profile.out
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
BenchmarkExportConntrackConns-2   	      98	  16741680 ns/op	  788391 B/op	    4847 allocs/op
PASS
ok  	antrea.io/antrea/pkg/agent/flowexporter/exporter	7.665s
```

Allocations have gone down by 54% and allocated space by 30%.